feat: class key event, started creating hud
